attackNavigator/navigator-server/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
Django==3.1.13
django-cors-headers==3.8.0
django-environ==0.8.0
django-sslserver==0.22
djangorestframework==3.12.4
et-xmlfile==1.0.1
filelock==3.0.12
fire==0.3.1
idna==2.10
jdcal==1.4.1
numpy==1.19.5
openpyxl==3.0.6
pandas==1.2.1
pendulum==1.2.5
Pillow==8.1.0
psycopg2==2.8.6
pyattck==3.0.1
pyfiglet==0.8.post1
python-dateutil==2.8.1
pytz==2020.5
pytzdata==2020.1
PyYAML==5.4.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.23
sqlparse==0.4.1
termcolor==1.1.0
tzlocal==2.1
urllib3==1.26.2
virtualenv==20.4.2
xlrd==2.0.1
